+++ dist-extension/service-worker-loader.js
@@ -1,1 +1,1 @@
-import './assets/background.ts-eJiyeCMk.js';
+import './assets/background.ts-Bmb_96WS.js';